# Use strstr() to increase performance