@Watson1978 2022-04-10
```ruby
require 'oj'

json =<<-EOF
{
  /* aa\0 aa */
  "name": "foobarbaz"
}
EOF
Oj.load(json)
```

Hmm, the behavior has been changed when \0 exists in a comment...